Move `CellGrid` to `typst-library/src/layout/cellgrid.rs`.
