+++ crates/typst-layout/src/grid/lines.rs
@@ -1,8 +1,8 @@
 use std::sync::Arc;
 
 use typst_library::foundations::{AlternativeFold, Fold};
+use typst_library::layout::cellgrid::{CellGrid, Line, Repeatable};
 use typst_library::layout::Abs;
-use typst_library::model::grid::{CellGrid, Line, Repeatable};
 use typst_library::visualize::Stroke;
 
 use super::RowPiece;
@@ -558,13 +558,13 @@ pub fn hline_stroke_at_column(
 
 #[cfg(test)]
 mod test {
+    use std::num::NonZeroUsize;
     use typst_library::foundations::Content;
     use typst_library::introspection::Locator;
     use typst_library::layout::{Axes, Sides, Sizing};
+    use typst_library::layout::cellgrid::{Entry, Cell, LinePosition};
     use typst_utils::NonZeroExt;
 
-    use super::super::cells::Entry;
-    use super::super::Cell;
     use super::*;
 
     fn sample_cell() -> Cell<'static> {
