+++ crates/typst-library/src/layout/mod.rs
@@ -4,6 +4,8 @@ mod abs;
 mod align;
 mod angle;
 mod axes;
+// TODO!
+pub mod cellgrid;
 mod columns;
 mod container;
 mod corners;
